#6 Cropping images to 2:1 aspect ratio
Cropping images to give them a uniform height so the grid looks nice. I
cropped at 2:1 to accommodate the images with text already on them. I
think we should consider not having text on the images (aside from the
logo) and letting the headlines speak for themselves (we could also
move the headlines over the image at that point).