Added standard footer with social media buttons
Adding a footer to the page with social media icons and links to the
rest of the website.
Reusing styles from
http://lendingcircles.force.com/pmtx/dn8n__SiteDonation?id=a81i0000000PA
sD

I made a few tweaks to center the logo and the address information
which looked broken to me.

If the MAF has that code hosted anywhere, I would be happy to fix it
there as well.

Fixes #52